Tweaks to the tribute
I changed the Δ to a 🖖 to better communicate the stories I'm referencing with it. (The starfleet logo has been called a delta, which is why I used it initially, forgetting that there's a better emoji to convey Trek.)

I changed the 🎼 to a 𝄞 because it shows up better printed on paper. In an astonishing coincidence, this post fits perfectly on letter paper. The only tweak I had to make was using Courier New rather than Arial for the tribute itself; the default 1" margins and size 11 font worked perfectly. The first copy I printed omitted the final paren; I don't think that means anything but it's noted.