Dear God, I hate having a messiah complex
I don't want a war, and I feel the accidents of my birth and life could put my words in the right ears. Are these the right ones? Will they stop a war that doesn't need to happen, start one that does, start one that doesn't, or stop one that does?

I am ignorant of so much. I haven't finished reading the Books, and haven't even touched the commentaries. I don't know the language, the meanings are lost on me, if I'm not simply wrong. I am a fool, a jester, in so many ways not a straight man. My whole life, I have been asked if I am joking, and not known the answer.

But if I know I have done all I can, maybe I can rest in peace without having to literally die.